+++ packages/powersync-attachments/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @journeyapps/powersync-attachments
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies [8cc1337]
+  - @journeyapps/powersync-sdk-common@1.5.1
+
 ## 2.0.1
 
 ### Patch Changes